+++ src/app/elvira/components/pdf-viewer/pdf-viewer.component.ts
@@ -3,10 +3,7 @@ import { pdfDefaultOptions } from 'ngx-extended-pdf-viewer';
 import { ActivatedRoute, Router } from '@angular/router';
 import { DisposableComponent } from 'src/app/common/components/disposable.component';
 import { AcquisitionService } from 'src/app/services/acquisition.service';
-import {
-  UserAcquisition,
-  UserAcquisitionId,
-} from 'src/app/types/acquisition.types';
+import { UserAcquisitionId } from 'src/app/types/acquisition.types';
 import { EntryService } from 'src/app/services/entry.service';
 import { TranslocoService } from '@ngneat/transloco';
 import { NotificationService } from 'src/app/services/general/notification.service';
@@ -63,7 +60,6 @@ export class PdfViewerComponent extends DisposableComponent implements OnInit {
   public base64Loaded: boolean = false; // used in html
   public citation: string | null; // send to wraper
   public acquisitionId: string; // send to wraper
-  userAcquisitionId: string;
 
   constructor(
     private readonly route: ActivatedRoute,
@@ -83,60 +79,41 @@ export class PdfViewerComponent extends DisposableComponent implements OnInit {
    * Than create user acquisition and get user acquisiton id
    * With user acquisition id get base64 file
    */
-  async ngOnInit(): Promise<void> {
-    const entry_id = this.route.snapshot.paramMap.get('entry_id');
+  ngOnInit(): void {
+    const entry_id = this.route.snapshot.paramMap.get('entry_id'); // get entry id
 
-    // Get details of given entry id
-    await this.entryService
-      .getEntryDetail(entry_id)
-      .toPromise()
-      .then(async (entryDetail) => {
+    try {
+      // Get details of given entry id
+      this.entryService.getEntryDetail(entry_id).subscribe((entryDetail) => {
         // set info from entry
         this.citation = entryDetail.response.citation
           ? entryDetail.response.citation
           : null;
         this.acquisitionId = entryDetail.response.acquisitions[0].id;
 
-        // Create user acquisition object
-        const userAcquisition: UserAcquisition = {
-          acquisition_id: this.acquisitionId,
-          type: 'personal',
-        };
-
-        // Call BE and create user acquistion and get user acquistion id of created user acquisition
-        await this.acquisitionService
-          .createUserAcquisition(userAcquisition)
-          .toPromise()
-          .then((res: UserAcquisitionId) => {
-            // get id
-            this.userAcquisitionId = res.response.id;
+        // Create user acquistion and get user acquistion id of created user acquisition
+        this.acquisitionService
+          .createUserAcquisition({
+            acquisition_id: this.acquisitionId,
+            type: 'personal',
           })
-          .catch(() => {
-            // If something went wrong
-            const message = this.translocoService.translate(
-              'lazy.pdfViewer.somethingWentWrong'
-            );
-            this.notificationService.error(message);
-            this.router.navigateByUrl('elvira/home');
+          .subscribe((res: UserAcquisitionId) => {
+            // Get base64 file from BE
+            this.acquisitionService
+              .getUserAcquisition(res.response.id, 'base64')
+              .subscribe((data) => {
+                this.base64 = data.response.data;
+                this.base64Loaded = true;
+              });
           });
-      })
-      .catch(() => {
-        // If something went wrong
-        const message = this.translocoService.translate(
-          'lazy.pdfViewer.somethingWentWrong'
-        );
-        this.notificationService.error(message);
-        this.router.navigateByUrl('elvira/home');
       });
-
-    // Get base64 file from BE
-    if (this.userAcquisitionId) {
-      this.acquisitionService
-        .getUserAcquisition(this.userAcquisitionId, 'base64')
-        .subscribe((data) => {
-          this.base64 = data.response.data;
-          this.base64Loaded = true;
-        });
+    } catch (error) {
+      // If something went wrong
+      const message = this.translocoService.translate(
+        'lazy.pdfViewer.somethingWentWrong'
+      );
+      this.notificationService.error(message);
+      this.router.navigateByUrl('elvira/home');
     }
   }
 }